+++ NEWS.md
@@ -1,3 +1,5 @@
+- check_names can now check names are unique
+- only check unique in check_named when unique = TRUE
 - check_colnames and check_names now accept (col)names = character(0)
 
 # checkr 0.0.1